+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - hel-1057: Ajout d'event Matomo lors de la consultation de la console admin
 - hel-1054: Ajout d'event Matomo lors de la consultation de la cartographie
 - hel-1053: Ajout d'event Matomo lors de la recherche simple
+- hel-1052: [Matomo] Indicateur Nb d'utilisateurs actifs/inactifs
 
 ## [1.8.10](https://github.com/DNUM-SocialGouv/helios/compare/release/1.8.9...release/1.8.10) (2026-06-15)